test: fix test

+++ internal/pkg/metadb/blobref/chunkref/chunkref_test.go
@@ -27,8 +27,8 @@ func TestChunkRef_New(t *testing.T) {
 	c := New(blobuuid, 42, 123450)
 	assert.NotEqual(t, uuid.Nil, c.Key)
 	assert.Equal(t, blobuuid, c.BlobRef)
-	assert.Equal(t, 42, c.Number)
-	assert.Equal(t, 123450, c.Size)
+	assert.Equal(t, int32(42), c.Number)
+	assert.Equal(t, int32(123450), c.Size)
 }
 
 func TestChunkRef_ObjectPath(t *testing.T) {
